refactor(sanitize): move argument sanitization to a dedicated package
